channels: Tombstone 4.1.37
It should not be stabilized because of [1], fixed in d9f0019 [2],
which landed after 4.1.37:

  $ oc adm release info --commits quay.io/openshift-release-dev/ocp-release:4.1.37-x86_64 | grep ' jenkins '
    jenkins                                       https://github.com/openshift/jenkins                                       76904bc459f27f8ea823d5c45a64a2748e782e4d
  $ git --no-pager log --oneline -3 origin/release-4.1
  d9f0019 (origin/release-4.1) Merge pull request #1033 from akram/4.1-nss-wrapper-2.204.2-and-plugins-updates
  1cd33d9 - Bump jenkins 2.204.2 and all plugins versions - Symlink to Dockerfile.localdev
  76904bc Merge pull request #982 from akram/4.1-bz-1764466-cve-2019-10431

I'm not clear on exactly when 4.1 broke, or whether releases older
than 4.1.37 are also affected.  However, there are no older 4.1
releases waiting for a prerelease -> fast promotion, so it's not
critical to pin down the regression timing (if this even was a
regression).

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1810393
[2]: https://github.com/openshift/jenkins/pull/1033#event-3114787745