channels: Add version-agnostic candidate, fast, and stable channels
For folks who want to flow from one 4.y to the next 4.(y+1) and onward
without having to manually bump their channel choice.  I'm seeding
these with the current content of the 4.3 channels.  I haven't gone
back to consider the *-4.1 and *-4.2 channels, because the
version-agnostic channels will not just be a blind union.  For
example, 4.1.20 is in stable-4.1, but not in stable-4.2.
blocked-edges/4.2.0.yaml blocks 4.1.20 -> 4.2.0, but without a more
careful audit than I've done, I'm not sure if all unsafe edges are
represented in blocked-edges (vs. just relying on the membership in
the channel files to avoid the unsafe edges).