4.4: Plan previous for 4.4.0
We don't want anything before 4.3.17, because [1] was fixed in 4.3.17
and the cluster-version operator doesn't yet use Upgradeable=False to
remove minor-version bumps from available updates.  So if folks on
earlier releases had DefaultSecurityContextConstraints_Mutated and saw
a 4.4.0 available update, they might choose it and then have the CVO
tell them "sorry, no minor bumps for you".  That's not great, so just
drop the update sources to force folks onto 4.3.17 or later to get the
condition cleared.

Also, the CRI-O / OCICNI context bug [2] was fixed in 4.3.18, so block
4.3.17 -> 4.4.0 as well.

Include a few future 4.3.z, in case we have a gap before 4.4.1 but
keep rolling with 4.3.z and want to continue to provide those folks
with a 4.3 -> 4.4 route.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1827337
[2]: https://bugzilla.redhat.com/show_bug.cgi?id=1828104